chore(assets): Move to folder
