+++ src/ocf_dataservices/definitions.py
@@ -1,17 +1,15 @@
 import dagster as dg
 
 from ocf_dataservices.defs.assets import (
-        l0_mars_ecmwf_ens_uk_v1,
-        l1_dynamical_ecmwf_ens_uk_v1,
-        l1_mars_ecmwf_ens_uk_v1,
-)
-from ocf_dataservices.defs.ecmwf_live import (
         ecmwf_live_s3_sensor,
         l0_ecmwf_live_local_v1,
         l0_ecmwf_live_s3_v1,
+        l0_mars_ecmwf_ens_uk_v1,
+        l1_dynamical_ecmwf_ens_uk_v1,
         l1_ecmwf_live_india_v1,
         l1_ecmwf_live_nl_v1,
         l1_ecmwf_live_uk_v1,
+        l1_mars_ecmwf_ens_uk_v1,
 )
 from ocf_dataservices.resources.iomanager_raw_file import (
         RawFileIOManager,
