+++ src/ocf_dataservices/defs/assets/__init__.py
@@ -1,0 +1,22 @@
+from ocf_dataservices.defs.assets.dynamical import l1_dynamical_ecmwf_ens_uk_v1
+from ocf_dataservices.defs.assets.ecmwf_live import (
+    ecmwf_live_s3_sensor,
+    l0_ecmwf_live_local_v1,
+    l0_ecmwf_live_s3_v1,
+    l1_ecmwf_live_india_v1,
+    l1_ecmwf_live_nl_v1,
+    l1_ecmwf_live_uk_v1,
+)
+from ocf_dataservices.defs.assets.ecmwf_mars import l0_mars_ecmwf_ens_uk_v1, l1_mars_ecmwf_ens_uk_v1
+
+__all__ = [
+    "ecmwf_live_s3_sensor",
+    "l0_ecmwf_live_local_v1",
+    "l0_ecmwf_live_s3_v1",
+    "l0_mars_ecmwf_ens_uk_v1",
+    "l1_dynamical_ecmwf_ens_uk_v1",
+    "l1_ecmwf_live_india_v1",
+    "l1_ecmwf_live_nl_v1",
+    "l1_ecmwf_live_uk_v1",
+    "l1_mars_ecmwf_ens_uk_v1",
+]
